server/face-recognition-master/requirements.txt
numpy
Pillow
dataset
facenet_pytorch
Flask
flask_restplus
joblib
matplotlib
seaborn
scikit_learn
torch
torchvision
Werkzeug
opencv-python
uWSGI
